add option to toggle fully integral check
